todo/requirements.txt
fastapi
uvicorn
mysql_connector_python
pydantic